+++ app/models/folio/file.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class Folio::File < Folio::ApplicationRecord
+  include Folio::HasHashId
   include Folio::Taggable
   include Folio::SanitizeFilename
   include Folio::MimeTypeDetection
@@ -46,6 +47,10 @@ class Folio::File < Folio::ApplicationRecord
     }
   end
 
+  def self.hash_id_additional_classes
+    [Folio::PrivateAttachment]
+  end
+
   private
 
     def touch_placements
@@ -76,6 +81,7 @@ end
 #  mime_type       :string(255)
 #  additional_data :json
 #  file_metadata   :json
+#  hash_id         :string
 #
 # Indexes
 #