+++ app/models/folio/image.rb
@@ -57,6 +57,7 @@ end
 #  mime_type       :string(255)
 #  additional_data :json
 #  file_metadata   :json
+#  hash_id         :string
 #
 # Indexes
 #